cli: rename packager to builder
